# docs/SWAGGER.json
{"openapi":"3.0.0","paths":{"/api/v1/scrapper/decathlon/{section}/{category}":{"get":{"operationId":"ScrapperController_getCatalogData","parameters":[{"name":"section","required":true,"in":"path","schema":{"type":"string"}},{"name":"category","required":true,"in":"path","schema":{"type":"string"}},{"name":"from","required":false,"in":"path","schema":{"type":"number"}},{"name":"size","required":false,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"records":{"type":"array","items":{"$ref":"#/components/schemas/Product"}},"totalSize":{"type":"number","example":1}}}}}}},"tags":["ScrapperService"]}},"/api/v1/scrapper/decathlon/{section}/{category}/{id}":{"get":{"operationId":"ScrapperController_getProductData","parameters":[{"name":"section","required":true,"in":"path","schema":{"type":"string"}},{"name":"category","required":true,"in":"path","schema":{"type":"string"}},{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"record":{"$ref":"#/components/schemas/Product"}}}}}}},"tags":["ScrapperService"]}},"/api/v1/catalog/decathlon/hierarchy":{"get":{"operationId":"CatalogController_getCatalogData","parameters":[],"responses":{"200":{"description":""}}}}},"info":{"title":"POSBC - Middleware","description":"POSB2C Middleware API Documentation","version":"0.0.1","contact":{}},"tags":[],"servers":[{"url":"http://localhost:3000","description":"Localhost"},{"url":"https://pos-b2c-d94e95f73222.herokuapp.com","description":"Dev Server"}],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"Product":{"type":"object","properties":{"title":{"type":"string","example":"Pantalon de trek montagne résistant Homme - MT500"},"brand":{"type":"string","example":"FORCLAZ"},"color":{"type":"string","example":"gris noir"},"reference":{"type":"string","example":"8853731"},"link":{"type":"string","example":"https://www.decathlon.fr/p/pantalon-de-trek-montagne-resistant-homme-mt500/_/R-p-351265?mc=8853731&c=gris_noir"},"currency":{"type":"string","example":"€"},"price":{"type":"object","example":36},"withDiscount":{"type":"boolean","example":true},"delivery":{"type":"string","example":"Livraison en 48h"},"beforeDiscount":{"type":"object","example":40},"discountAmount":{"type":"object","example":4},"discountDateInfo":{"type":"string","example":"Offre valable jusqu’au 31/12/2021"}},"required":["title","brand","color","reference","link","currency","price","withDiscount","delivery","beforeDiscount","discountAmount","discountDateInfo"]}}}}